products get added to cart, including multiples adds
