+++ frontend/src/components/XProduct.tsx
@@ -1,9 +1,6 @@
 "use client";
 
-import { storeCart } from "@/store/storeCart";
-import { CartProduct } from "@/types/cart";
 import { Product } from "@/types/types";
-import React, { useState } from "react";
 import Image from "next/image";
 import AddToCart from "./XAddToCart";
 
@@ -12,36 +9,6 @@ interface XProductProps {
 }
 
 const XProduct = ({ product }: XProductProps) => {
-  const [selectedQuantities, setSelectedQuantities] = useState<{ [key: string]: number }>({});
-  const { items, addItem } = storeCart();
-
-
-  const updateQuantity = (productId: string, delta: number) => {
-    setSelectedQuantities(prev => {
-      const currentQty = prev[productId] || 0;
-      const newQty = Math.max(0, Math.min(currentQty + delta, product.inventory_count || 0));
-      return { ...prev, [productId]: newQty };
-    });
-  };
-
-  const handleAddToCart = () => {
-    const item: CartProduct | undefined = items.find((item) => item.id === product.id);
-    if (item) {
-      updateQuantity(product.id, item.quantity + 1);
-    } else {
-      const cartProduct: CartProduct = {
-        id: product.id,
-        name: product.name,
-        description: product.description,
-        price: product.price,
-        imageUrl: product.imageUrl,
-        country: product.country,
-        inventory_count: product.inventory_count,
-        quantity: 1,
-      };
-      addItem(cartProduct);
-    }
-  };
 
   return (
     <div>
@@ -56,7 +23,7 @@ const XProduct = ({ product }: XProductProps) => {
       </div>
       <div className="text-2xl font-bold my-4">Price: ${(product.price / 100).toFixed(2)}</div>
       <div className="my-4">
-        <AddToCart product={product} selectedQuantities={selectedQuantities} updateQuantity={updateQuantity} />
+        <AddToCart product={product} />
         <div className="text-md my-4"><strong>Status:</strong> {product.inventory_count > 0 ? 'In Stock' : 'Out of Stock'}</div>
       </div>
       <div className="text-md my-4"><strong>Country of Origin:</strong> {product.country}</div>
