+++ frontend/src/store/storeCart.ts
@@ -5,13 +5,13 @@ export const storeCart = create<CartStore>((set) => ({
   items: [],
   total: 0,
 
-  addItem: (product) =>
+  addItem: (product, quantity=1) =>
     set((state) => {
       const existingItem = state.items.find((item) => item.id === product.id);
 
       if (existingItem) {
         const updatedItems = state.items.map((item) =>
-          item.id === product.id ? { ...item, quantity: (item.quantity || 1) + 1 } : item
+          item.id === product.id ? { ...item, quantity: (item.quantity || 1) + quantity } : item
         );
         return {
           items: updatedItems,
@@ -19,7 +19,7 @@ export const storeCart = create<CartStore>((set) => ({
         };
       }
 
-      const newItems = [...state.items, { ...product, quantity: 1 }];
+      const newItems = [...state.items, { ...product, quantity: quantity }];
       return {
         items: newItems,
         total: calculateTotal(newItems),
@@ -34,8 +34,8 @@ export const storeCart = create<CartStore>((set) => ({
         total: calculateTotal(newItems),
       };
     }),
-
-  updateQuantity: (productId, quantity) =>
+  
+    updateQuantity: (productId, quantity) =>
     set((state) => {
       if (quantity < 1) return state;
